Move harmony metadata into it's own file in the desync report
It can easily generate over 150kB of data. Moving it out to a separate
file makes sense content-wise and makes it easier to browse the log file
