+++ Source/Client/Desyncs/LogGenerator.cs
@@ -1,12 +1,7 @@
 using System;
-using System.Diagnostics;
 using System.IO;
 using System.Linq;
-using System.Reflection;
-using System.Text;
 using System.Text.RegularExpressions;
-using System.Xml.Linq;
-using HarmonyLib;
 using UnityEngine;
 using Verse;
 
@@ -15,7 +10,7 @@ namespace Multiplayer.Client
     /// <summary>
     /// Collects the game logs and loaded mods to be included with desync file.
     /// </summary>
-    public class LogGenerator
+    public static class LogGenerator
     {
         private const int MaxLogLineCount = 10000;
 
@@ -38,12 +33,7 @@ namespace Multiplayer.Client
                 }
                 else logSection = "Could not find the log";
 
-                var collatedData = string.Concat(MakeLogTimestamp(),
-                    ListActiveMods(), "\n",
-                    ListHarmonyPatches(), "\n",
-                    logSection);
-
-                return collatedData;
+                return logSection;
             }
             catch
             {
@@ -164,12 +154,7 @@ namespace Multiplayer.Client
             var fileContents = File.ReadAllText(tempPath);
             File.Delete(tempPath);
 
-            return "Log file contents:\n" + fileContents;
-        }
-
-        private static string MakeLogTimestamp()
-        {
-            return string.Concat("Log generated on ", DateTime.Now.ToLongDateString(), ", ", DateTime.Now.ToLongTimeString(), "\n");
+            return fileContents;
         }
 
         private static string RedactString(string original, string redactStart, string redactEnd, string replacement)
@@ -186,81 +171,6 @@ namespace Multiplayer.Client
             result += logTail;
 
             return result;
-        }
-
-        public static string ListHarmonyPatches()
-        {
-            var patchListing = DescribeAllPatchedMethods();
-
-            return string.Concat("Active Harmony patches:\n",
-                patchListing,
-                patchListing.EndsWith("\n") ? "" : "\n",
-                HarmonyUtil.DescribeHarmonyVersions(), "\n");
-        }
-
-        private static string ListActiveMods()
-        {
-            var builder = new StringBuilder();
-            builder.Append("Loaded mods:\n");
-            foreach (var modContentPack in LoadedModManager.RunningMods)
-            {
-                builder.AppendFormat("{0}({1})", modContentPack.Name, modContentPack.PackageIdPlayerFacing);
-                TryAppendOverrideVersion(builder, modContentPack);
-                TryAppendManifestVersion(builder, modContentPack);
-                builder.Append(": ");
-                var firstAssembly = true;
-                var anyAssemblies = false;
-                foreach (var loadedAssembly in modContentPack.assemblies.loadedAssemblies)
-                {
-                    if (!firstAssembly)
-                    {
-                        builder.Append(", ");
-                    }
-                    firstAssembly = false;
-                    builder.Append(loadedAssembly.GetName().Name);
-
-                    var (version, fileVersion) = ReadModAssembly(loadedAssembly, modContentPack);
-
-                    if (version != null)
-                    {
-                        if (fileVersion == null) builder.AppendFormat("({0} [no file version])", version);
-                        else if (version == fileVersion) builder.AppendFormat("({0})", version);
-                        else builder.AppendFormat("(av: {0} fv:{1})", version, fileVersion);
-                    }
-
-                    anyAssemblies = true;
-                }
-                if (!anyAssemblies)
-                {
-                    builder.Append("(no assemblies)");
-                }
-                builder.Append("\n");
-            }
-            return builder.ToString();
-        }
-
-        private static void TryAppendOverrideVersion(StringBuilder builder, ModContentPack pack)
-        {
-            var filePath = Path.Combine(pack.RootDir, Path.Combine("About", "Version.xml"));
-            if (!File.Exists(filePath)) return;
-
-            var doc = XDocument.Load(filePath);
-
-            var overrideVersionElement = doc.Root?.Element("overrideVersion");
-            if (overrideVersionElement != null)
-                builder.AppendFormat("[ov:{0}]", overrideVersionElement.Value);
-        }
-
-        private static void TryAppendManifestVersion(StringBuilder builder, ModContentPack pack)
-        {
-            if (pack == null) return;
-
-            var filePath = Path.Combine(pack.RootDir, Path.Combine("About", "Manifest.xml"));
-            if (!File.Exists(filePath)) return;
-
-            var doc = XDocument.Load(filePath);
-            var versionElement = doc.Root?.Element("version") ?? doc.Root?.Element("Version");
-            if (versionElement != null) builder.AppendFormat("[mv:{0}]", versionElement.Value);
         }
 
         /// <summary>
@@ -285,21 +195,6 @@ namespace Multiplayer.Client
             };
         }
 
-        /// <summary>
-        /// Produces a human-readable list of all methods patched by all Harmony instances and their respective patches.
-        /// </summary>
-        private static string DescribeAllPatchedMethods()
-        {
-            try
-            {
-                return HarmonyUtil.DescribePatchedMethodsList(Harmony.GetAllPatchedMethods());
-            }
-            catch (Exception e)
-            {
-                return "Could not retrieve patched methods from the Harmony library:\n" + e;
-            }
-        }
-
         private static string TryGetCommandLineOptionValue(string key)
         {
             const string keyPrefix = "-";
@@ -317,42 +212,6 @@ namespace Multiplayer.Client
                 }
             }
             return null;
-        }
-
-        /// <summary>
-        /// Reads assembly version information for a mod assembly.
-        /// Mod assemblies require special treatment, since they are loaded from byte arrays and their <see cref="Assembly.Location"/> is null.
-        /// </summary>
-        /// <param name="assembly">The assembly to read</param>
-        /// <param name="contentPack">The content pack the assembly was loaded from</param>
-        public static (string assemblyVersion, string assemblyFileVersion) ReadModAssembly(Assembly assembly, ModContentPack contentPack)
-        {
-            static string ToSemanticString(Version v)
-            {
-                // System.Version parts: Major.Minor.Build.Revision
-                return v.Build < 0
-                    ? $"{v.ToString(2)}.0"
-                    : v.ToString(v.Revision <= 0 ? 3 : 4);
-            }
-
-            if (assembly == null) return (null, null);
-            const string assembliesFolderName = "Assemblies";
-
-            var expectedAssemblyFileName = $"{assembly.GetName().Name}.dll";
-            var modAssemblyFolderFiles = ModContentPack.GetAllFilesForMod(contentPack, assembliesFolderName);
-            var fileHandle = modAssemblyFolderFiles.Values.FirstOrDefault(f => f.Name == expectedAssemblyFileName);
-
-            try
-            {
-                var assemblyFilePath = fileHandle?.FullName ?? assembly.Location;
-                var fileInfo = FileVersionInfo.GetVersionInfo(assemblyFilePath);
-
-                return (ToSemanticString(assembly.GetName().Version), ToSemanticString(new Version(fileInfo.FileVersion)));
-            }
-            catch (Exception)
-            {
-                return (ToSemanticString(assembly.GetName().Version), null);
-            }
         }
     }
 }
